requirements.txt
python-dotenv
openai
openai-agents
requests
pypdf
gradio
pydantic
sendgrid